feat:매트릭스 제외 map사용 부문에서 key 설정 매트릭스의 경우 해결방법을 찾아 차후 수정
