+++ client/src/components/calendar/CalendarView/index.tsx
@@ -11,7 +11,11 @@ function CalendarView({ totalInOut, lang, year, month }: CalendarViewType): JSX.
       <MatrixView
         headers={getWeekDays(lang)}
         matrix={getDayMatrix(year, month)}
-        cell={(date, key) => <TableCell day={date} key={key} totalInOut={totalInOut} />}
+        cell={(date, key) => (
+          <>
+            <TableCell day={date} key={key} totalInOut={totalInOut} />
+          </>
+        )}
       />
     </>
   );
