--- client/src/container/Calendar/types.ts
@@ -1,4 +1,0 @@
-export type CalendarType = {
-  month: number;
-  year: number;
-};
