+++ client/src/libs/mockData.ts
@@ -1,0 +1,227 @@
+const mockTransactions = [
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-12-01',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-10-05',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-12-03',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-12-18',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 5000,
+    tradeAt: '2020-12-20',
+    description: 'user1의 가계부내역 82',
+    isIncome: false,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: false,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-12-21',
+    description: 'user1의 가계부내역 82',
+    isIncome: false,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: false,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 5000,
+    tradeAt: '2020-10-22',
+    description: 'user1의 가계부내역 82',
+    isIncome: false,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: false,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-10-23',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-12-27',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+  {
+    tid: 82,
+    amount: 10000,
+    tradeAt: '2020-12-30',
+    description: 'user1의 가계부내역 82',
+    isIncome: true,
+    uid: 1,
+    cid: 2,
+    pid: 3,
+    payment: {
+      pid: 3,
+      name: '농협은행',
+      uid: 1,
+    },
+    category: {
+      cid: 2,
+      name: '식비',
+      isIncome: true,
+      uid: 1,
+    },
+  },
+];
+
+const totalInOut: any = {
+  1: { totalIn: 150, totalOut: 5000 },
+  5: { totalIn: 10000, totalOut: 0 },
+  13: { totalIn: 10000, totalOut: 0 },
+  18: { totalIn: 50000, totalOut: 40000 },
+  20: { totalIn: 20000, totalOut: 25000 },
+  21: { totalIn: 0, totalOut: 5000 },
+  22: { totalIn: 10000, totalOut: 0 },
+  23: { totalIn: 0, totalOut: 5000 },
+  27: { totalIn: 10000, totalOut: 0 },
+  30: { totalIn: 0, totalOut: 10000 },
+};
+
+export { mockTransactions, totalInOut };
